+++ docs/guides/changelog.md
@@ -10,6 +10,7 @@ description: |-
 ## 0.1.6 (unreleased)
 
 - Add `ise_downloadable_acl` resource and data source
+- Add `ise_tacacs_command_set` resource and data source
 
 ## 0.1.5